feat(tree): support json, yaml to tree
