+++ crates/serde_hkx_features/src/lib.rs
@@ -2,7 +2,10 @@ pub mod convert;
 pub mod diff;
 pub mod dump;
 pub mod error;
-pub mod read_ext;
+pub mod fs;
+pub mod serde;
 pub mod tree;
 
-pub(crate) type ClassMap<'a> = indexmap::IndexMap<usize, havok_classes::Classes<'a>>;
+/// - key: class index(e.g `#0001`)
+/// - value: C++ Class
+pub type ClassMap<'a> = indexmap::IndexMap<usize, havok_classes::Classes<'a>>;
